% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{get_all_lake_data,Endpoint-method}
\alias{get_all_lake_data,Endpoint-method}
\title{Returns most recent data for all registered lakes}
\usage{
\S4method{get_all_lake_data}{Endpoint}(endpoint)
}
\arguments{
\item{endpoint}{an Endpoint object}
}
\description{
Returns most recent data for all registered lakes
}
