% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{get_all_lake_data,RequestHandler-method}
\alias{get_all_lake_data,RequestHandler-method}
\title{Returns most recent data for all lakes}
\usage{
\S4method{get_all_lake_data}{RequestHandler}(endpoint)
}
\arguments{
\item{endpoint}{a RequestHandler object}
}
\value{
a tibble of most recent data from all registered lakes
}
\description{
Returns most recent data for all lakes
}
