% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{get_all_lake_names,Endpoint-method}
\alias{get_all_lake_names,Endpoint-method}
\title{Returns the names of all lakes registered in the database}
\usage{
\S4method{get_all_lake_names}{Endpoint}(endpoint)
}
\description{
Returns the names of all lakes registered in the database
}
