% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{get_all_lake_names,RequestHandler-method}
\alias{get_all_lake_names,RequestHandler-method}
\title{Returns all lake names}
\usage{
\S4method{get_all_lake_names}{RequestHandler}(endpoint)
}
\arguments{
\item{endpoint}{a RequestHandler object}
}
\value{
a tibble containing lake names
}
\description{
Returns all lake names
}
