% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{get_lake_data,Endpoint,character-method}
\alias{get_lake_data,Endpoint,character-method}
\title{returns most recent data for a specified lake. The name should be all lowercase, use
underscores instead of spaces, and should not have a "Lake" suffix. e.g.,
"bde_maka_ska"}
\usage{
\S4method{get_lake_data}{Endpoint,character}(endpoint, lake_name)
}
\arguments{
\item{endpoint}{an Endpoint object}

\item{lake_name}{a string correctly specifying the name of a lake}
}
\description{
returns most recent data for a specified lake. The name should be all lowercase, use
underscores instead of spaces, and should not have a "Lake" suffix. e.g.,
"bde_maka_ska"
}
