% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{get_lake_history,RequestHandler,character-method}
\alias{get_lake_history,RequestHandler,character-method}
\title{Returns all lake data within a specified date range for a given lake.
The lake name should be all lowercase, use underscores instead of spaces, and
 should not have a "Lake" suffix. e.g., "bde_maka_ska"}
\usage{
\S4method{get_lake_history}{RequestHandler,character}(endpoint, lake_name, start_date, end_date = Sys.Date())
}
\arguments{
\item{endpoint}{an RequestHandler object}

\item{lake_name}{a string correctly specifying the name of a lake}

\item{start_date}{oldest date to include}

\item{end_date}{most recent date to include (defaults to today)}
}
\description{
Returns all lake data within a specified date range for a given lake.
The lake name should be all lowercase, use underscores instead of spaces, and
 should not have a "Lake" suffix. e.g., "bde_maka_ska"
}
