% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{request_handler}
\alias{request_handler}
\title{Returns a request handler}
\usage{
request_handler(type)
}
\arguments{
\item{type}{Determines if you are using a mock handler with mock data (for
testing) or a "real" handler with live data. Use "test" for the mock handler
and "live" for deployment}
}
\description{
Returns a request handler
}
