% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{fromJSON}
\alias{toJSON}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{jsonlite}{\code{\link[jsonlite]{fromJSON}}, \code{\link[jsonlite:fromJSON]{toJSON}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

